Remove "computing digest" sub-job.
It doesn't seem very helpful, and having a sub-job messes up progress
reporting when examining lots of files (each of which doesn't take very
long) as in the UI we just end up seeing the 0% of each sub-job one
after the other.

When we have a lot of sub jobs we really want to just report overall
progress, I think, especially if the sub jobs will last roughly the same
time.